+++ CBayesianSearch.h
@@ -31,7 +31,7 @@ class BayesianSearchModel {
         BayesianSearchModel() {}
         BayesianSearchModel(CCmpndKern& kernel, const CMatrix& param_bounds,
                             double observation_noise, double exp_bias,
-                            int init_samples, int rng_seed, int verbose, vector<CMatrix> grid=vector<CMatrix>()) {
+                            size_t init_samples, int rng_seed, int verbose, vector<CMatrix> grid=vector<CMatrix>()) {
             kern = kernel.clone();
             DIMENSIONMATCH(param_bounds.getCols() == 2);
             bounds = param_bounds;
@@ -99,9 +99,9 @@ class BayesianSearchModel {
             rng.seed(seed);
         }
 
-        int num_samples;
-        int x_dim;
-        int initial_samples;
+        size_t num_samples;
+        size_t x_dim;
+        size_t initial_samples;
         int seed;
         int verbosity;
         std::mt19937 rng;
